Lector de Atracciones

+++ src/tierra_Media/Producto.java
@@ -2,7 +2,7 @@ package tierra_Media;
 
 public interface Producto {
 	public Tipo_de_Atraccion getTipoDeAtraccion();
-	public int getCostoDeVisita();
+	public double getCostoDeVisita();
 
 	
 }
